@@ -1,6 +1,12 @@
 package murraco.controller;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 
 import lombok.RequiredArgsConstructor;
 import murraco.model.AppUser;
@@ -15,19 +21,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import io.swagger.annotations.Authorization;
 import murraco.dto.UserDataDTO;
 import murraco.dto.UserResponseDTO;
 import murraco.service.UserService;
 
 @RestController
 @RequestMapping("/users")
-@Api(tags = "users")
+@Tag(name = "users")
 @RequiredArgsConstructor
 public class UserController {
 
@@ -35,58 +35,61 @@ public class UserController {
   private final ModelMapper modelMapper;
 
   @PostMapping("/signin")
-  @ApiOperation(value = "${UserController.signin}")
-  @ApiResponses(value = {//
-      @ApiResponse(code = 400, message = "Something went wrong"), //
-      @ApiResponse(code = 422, message = "Invalid username/password supplied")})
-  public String login(//
-      @ApiParam("Username") @RequestParam String username, //
-      @ApiParam("Password") @RequestParam String password) {
+  @Operation(summary = "${UserController.signin}")
+  @ApiResponses(value = {
+         @ApiResponse(responseCode = "400", description = "Something went wrong"),
+         @ApiResponse(responseCode = "422", description = "Invalid username/password supplied")
+  })
+  public String login(
+      @Parameter(description = "Username") @RequestParam String username,
+      @Parameter(description = "Password") @RequestParam String password) {
     return userService.signin(username, password);
   }
 
   @PostMapping("/signup")
-  @ApiOperation(value = "${UserController.signup}")
-  @ApiResponses(value = {//
-      @ApiResponse(code = 400, message = "Something went wrong"), //
-      @ApiResponse(code = 403, message = "Access denied"), //
-      @ApiResponse(code = 422, message = "Username is already in use")})
-  public String signup(@ApiParam("Signup User") @RequestBody UserDataDTO user) {
+  @Operation(summary = "${UserController.signup}")
+  @ApiResponses(value = {
+         @ApiResponse(responseCode = "400", description = "Something went wrong"),
+         @ApiResponse(responseCode = "403", description = "Access denied"),
+         @ApiResponse(responseCode = "422", description = "Username is already in use")
+  })
+  public String signup(@Parameter(description = "Signup User") @RequestBody UserDataDTO user) {
     return userService.signup(modelMapper.map(user, AppUser.class));
   }
 
   @DeleteMapping(value = "/{username}")
   @PreAuthorize("hasRole('ROLE_ADMIN')")
-  @ApiOperation(value = "${UserController.delete}", authorizations = { @Authorization(value="apiKey") })
-  @ApiResponses(value = {//
-      @ApiResponse(code = 400, message = "Something went wrong"), //
-      @ApiResponse(code = 403, message = "Access denied"), //
-      @ApiResponse(code = 404, message = "The user doesn't exist"), //
-      @ApiResponse(code = 500, message = "Expired or invalid JWT token")})
-  public String delete(@ApiParam("Username") @PathVariable String username) {
+  @Operation(summary = "${UserController.delete}", security = @SecurityRequirement(name = "apiKey"))
+  @ApiResponses(value = {
+         @ApiResponse(responseCode = "400", description = "Something went wrong"),
+         @ApiResponse(responseCode = "403", description = "Access denied"),
+         @ApiResponse(responseCode = "404", description = "The user doesn't exist"),
+         @ApiResponse(responseCode = "500", description = "Expired or invalid JWT token")
+  })
+  public String delete(@Parameter(description = "Username") @PathVariable String username) {
     userService.delete(username);
     return username;
   }
 
   @GetMapping(value = "/{username}")
   @PreAuthorize("hasRole('ROLE_ADMIN')")
-  @ApiOperation(value = "${UserController.search}", response = UserResponseDTO.class, authorizations = { @Authorization(value="apiKey") })
-  @ApiResponses(value = {//
-      @ApiResponse(code = 400, message = "Something went wrong"), //
-      @ApiResponse(code = 403, message = "Access denied"), //
-      @ApiResponse(code = 404, message = "The user doesn't exist"), //
-      @ApiResponse(code = 500, message = "Expired or invalid JWT token")})
-  public UserResponseDTO search(@ApiParam("Username") @PathVariable String username) {
+  @Operation(summary = "${UserController.search}", responses = {
+         @ApiResponse(responseCode = "400", description = "Something went wrong"),
+         @ApiResponse(responseCode = "403", description = "Access denied"),
+         @ApiResponse(responseCode = "404", description = "The user doesn't exist"),
+         @ApiResponse(responseCode = "500", description = "Expired or invalid JWT token")
+  }, security = @SecurityRequirement(name = "apiKey"))
+  public UserResponseDTO search(@Parameter(description = "Username") @PathVariable String username) {
     return modelMapper.map(userService.search(username), UserResponseDTO.class);
   }
 
   @GetMapping(value = "/me")
   @PreAuthorize("hasRole('ROLE_ADMIN') or hasRole('ROLE_CLIENT')")
-  @ApiOperation(value = "${UserController.me}", response = UserResponseDTO.class, authorizations = { @Authorization(value="apiKey") })
-  @ApiResponses(value = {//
-      @ApiResponse(code = 400, message = "Something went wrong"), //
-      @ApiResponse(code = 403, message = "Access denied"), //
-      @ApiResponse(code = 500, message = "Expired or invalid JWT token")})
+  @Operation(summary = "${UserController.me}", responses = {
+         @ApiResponse(responseCode = "400", description = "Something went wrong"),
+         @ApiResponse(responseCode = "403", description = "Access denied"),
+         @ApiResponse(responseCode = "500", description = "Expired or invalid JWT token")
+  }, security = @SecurityRequirement(name = "apiKey"))
   public UserResponseDTO whoami(HttpServletRequest req) {
     return modelMapper.map(userService.whoami(req), UserResponseDTO.class);
   }
